Add `ets:new/2` option to disable decentralized counters
This commit adds an option that can be passed to the `ets:new/2`
function to create an ETS table with the `write_concurrency` option
enabled that uses centralized counters to keep track of the number of
items in the table and the memory consumption of the table. Tables
with the `write_concurrency` option enabled use decentralized counters
by default.

The functions `ets:info/1` and `ets:info/2` (with `size` or `memory`
as the the second parameter) get substantially slower and less
scalable in a table that use decentralized counters compared to a
table that use centralized counters so this option is useful for
applications that frequently call `ets:info/1` or `ets:info/2`.