ETS `ordered_set`: Add gravity to CA tree implementation
This commit only affects the implementation of ETS `ordered_set`
tables with the write_concurrency option enabled. Such tables are
implemented with a data structure that is called the contention
adapting search tree (CA tree). With this commit, about 1 in 4096 read
unlock calls for base node locks will trigger a join in one randomly
selected base node.

Before this commit, operations that only acquired looks in read-mode
never triggered any contention adaptation (splits or joins of base
nodes that cause the CA tree to use more or less fine-grained
locking). Therefore, the CA tree could get stuck in a sub-optimal
state in certain scenarios. This could happen, for example, when a
CA tree is first populated with parallel inserts (which will cause
splits of base nodes) and then only read-only operations are applied
to the data structure. Benchmark results from the
`ets_SUITE:lookup_catree_par_vs_seq_init_benchmark/0` benchmark
function (which is included in this commit) shows that this change can
improve the throughput of the CA tree in the scenario described above.