Use special environment variable to change default ROOTDIR
Using the environment variable ROOTDIR to change to a custom root dir
for the Erlang/OTP installation is not a good idea because this
variable might sometimes be set to something incorrect. This commit
changes so that the the user can change the root directory by setting
the environment variable CUSTOM_ERL_ROOTDIR instead.